+++ PSWorkItem.psm1
@@ -1,4 +1,3 @@
-
 # used for culture debugging
 # write-host "Importing with culture $(Get-Culture)"